Set up bid automaton tests, remove obsolete test
