+++ pkg/core/consensus/stakeautomaton/stakeautomaton_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus"
 	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus/stakeautomaton"
-	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus/user"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire/message"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire/topics"
 	"github.com/dusk-network/dusk-blockchain/pkg/util/nativeutils/eventbus"
@@ -22,7 +21,7 @@ const pass = "password"
 // Test that the maintainer will properly send new stake transactions, when
 // one is about to expire, or if none exist.
 func TestMaintainStakes(t *testing.T) {
-	bus, rb, p := setupMaintainerTest(t)
+	bus, rb := setupMaintainerTest(t)
 	defer func() {
 		_ = os.Remove("wallet.dat")
 		_ = os.RemoveAll("walletDB")
@@ -32,7 +31,7 @@ func TestMaintainStakes(t *testing.T) {
 	go catchStakeRequest(t, rb, c)
 
 	// Send round update, to start the maintainer.
-	ru := consensus.MockRoundUpdate(1, p)
+	ru := consensus.MockRoundUpdate(1, nil)
 	ruMsg := message.New(topics.RoundUpdate, ru)
 	bus.Publish(topics.RoundUpdate, ruMsg)
 
@@ -40,7 +39,7 @@ func TestMaintainStakes(t *testing.T) {
 	<-c
 
 	// Then, send a round update close after. No stake request should be sent
-	ru = consensus.MockRoundUpdate(2, p)
+	ru = consensus.MockRoundUpdate(2, nil)
 	ruMsg = message.New(topics.RoundUpdate, ru)
 	bus.Publish(topics.RoundUpdate, ruMsg)
 
@@ -53,8 +52,8 @@ func TestMaintainStakes(t *testing.T) {
 		// success
 	}
 
-	// Send another round update that is within the 'offset', to trigger sending a new pair of txs
-	ru = consensus.MockRoundUpdate(950, p)
+	// Send another round update that is within the 'offset', to trigger sending a new tx
+	ru = consensus.MockRoundUpdate(950, nil)
 	ruMsg = message.New(topics.RoundUpdate, ru)
 	bus.Publish(topics.RoundUpdate, ruMsg)
 
@@ -62,45 +61,7 @@ func TestMaintainStakes(t *testing.T) {
 	<-c
 }
 
-// Ensure the maintainer does not keep sending bids and stakes until they are included.
-func TestSendOnce(t *testing.T) {
-	bus, rb, p := setupMaintainerTest(t)
-	defer func() {
-		_ = os.Remove("wallet.dat")
-		_ = os.RemoveAll("walletDB")
-	}()
-
-	c := make(chan struct{}, 1)
-	go catchStakeRequest(t, rb, c)
-
-	time.Sleep(100 * time.Millisecond)
-
-	// Send round update, to start the maintainer.
-	ru := consensus.MockRoundUpdate(1, p)
-	ruMsg := message.New(topics.RoundUpdate, ru)
-	bus.Publish(topics.RoundUpdate, ruMsg)
-
-	// Ensure stake request is sent
-	<-c
-
-	go catchStakeRequest(t, rb, c)
-
-	time.Sleep(100 * time.Millisecond)
-
-	// Update round
-	ru = consensus.MockRoundUpdate(2, p)
-	ruMsg = message.New(topics.RoundUpdate, ru)
-	bus.Publish(topics.RoundUpdate, ruMsg)
-
-	select {
-	case <-c:
-		t.Fatal("was not supposed to get another tx in txChan")
-	case <-time.After(2 * time.Second):
-		// success
-	}
-}
-
-func setupMaintainerTest(t *testing.T) (*eventbus.EventBus, *rpcbus.RPCBus, *user.Provisioners) {
+func setupMaintainerTest(t *testing.T) (*eventbus.EventBus, *rpcbus.RPCBus) {
 	bus := eventbus.New()
 	rpcBus := rpcbus.New()
 
@@ -108,10 +69,7 @@ func setupMaintainerTest(t *testing.T) (*eventbus.EventBus, *rpcbus.RPCBus, *use
 	_, err := m.AutomateConsensusTxs(context.Background(), &node.EmptyRequest{})
 	require.Nil(t, err)
 
-	// Mock provisioners, and insert our wallet values
-	p, _ := consensus.MockProvisioners(10)
-
-	return bus, rpcBus, p
+	return bus, rpcBus
 }
 
 func catchStakeRequest(t *testing.T, rb *rpcbus.RPCBus, respChan chan struct{}) {
